+++ HISTORY.rst
@@ -3,10 +3,14 @@
 History
 -------
 
+0.1.1 (2017-06-22)
+++++++++++++++++++
+* add module entity with SimpleEntity.ner() 
 
 0.1.0 (2017-06-19)
 ++++++++++++++++++
 
 * initial PyPI release
-* support json, table(excel), web modules
+* add module json, table(excel), web for data manipulation
+* provide cli ui via misc.main_subtask
 * connect to travis CI